Update README.md
This project runs on a Unix Terminal. It provides a user-friendly interface and implements an online banking system. I do not provide any warranty for this project. Please use and modify this with your own consent.
How to run this project?
On terminal 1 -> gcc -o server server.c
				 then run as ./server
On terminal 2 -> gcc -o client client.c
				 then run as ./client
Admin Credentials:
username - admin
password - root

Initially, no database and files have been created. They will be created as and when necessary at certain steps.
Separate structures have been implemented for joint users and normal users for ease of implementation.
Socket programming has been used to communicate between client and server. All the transactions and database updations have been done on the server-side only. The client sends all the data given by the user to the server.

File locking has been implemented to prevent simultaneous read/write access.
At a time, only one instance of a customer can log in to the server.
In the case of a joint account, if one of the account holders has logged in, the other user of that joint account cannot log in to the server.
Necessary information has been stored in the database files which are separate for normal customers and joint account customers, i.e, Customer_Accounts, and Joint_Accounts. A separate temporary DB file has also been created to facilitate the deletion of records from the database.